+++ backend/common/settings/structs.go
@@ -28,8 +28,6 @@ type Server struct {
 	TLSCert            string      `json:"tlsCert"`
 	EnableThumbnails   bool        `json:"enableThumbnails"`
 	ResizePreview      bool        `json:"resizePreview"`
-	EnableExec         bool        `json:"enableExec"`
-	AuthHook           string      `json:"authHook"`
 	Port               int         `json:"port"`
 	BaseURL            string      `json:"baseURL"`
 	Logging            []LogConfig `json:"logging"`
@@ -40,21 +38,21 @@ type Server struct {
 	CacheDir           string      `json:"cacheDir"`
 	MaxArchiveSizeGB   int64       `json:"maxArchiveSize"`
 	// not exposed to config
-	SourceMap     map[string]Source `json:"-"` // uses realpath as key
-	NameToSource  map[string]Source `json:"-"` // uses name as key
-	DefaultSource Source            `json:"-"`
+	SourceMap     map[string]Source `json:"-" validate:"omitempty"` // uses realpath as key
+	NameToSource  map[string]Source `json:"-" validate:"omitempty"` // uses name as key
+	DefaultSource Source            `json:"-" validate:"omitempty"`
 }
 
 type Integrations struct {
-	OnlyOffice OnlyOffice `json:"office"`
-	Media      Media      `json:"media"`
+	OnlyOffice OnlyOffice `json:"office" validate:"omitempty"`
+	Media      Media      `json:"media" validate:"omitempty"`
 }
 
 // onlyoffice secret is stored in the local.json file
 // docker exec <containerID> /var/www/onlyoffice/documentserver/npm/json -f /etc/onlyoffice/documentserver/local.json 'services.CoAuthoring.secret.session.string'
 type OnlyOffice struct {
-	Url    string `json:"url"`
-	Secret string `json:"secret"`
+	Url    string `json:"url" validate:"required"`
+	Secret string `json:"secret" validate:"required"`
 }
 
 type Media struct {
@@ -70,8 +68,8 @@ type LogConfig struct {
 }
 
 type Source struct {
-	Path   string       `json:"path"` // can be relative, filesystem path
-	Name   string       `json:"name"` // display name
+	Path   string       `json:"path" validate:"required"` // file system path. (Can be relative)
+	Name   string       `json:"name"`                     // display name
 	Config SourceConfig `json:"config"`
 }
 
@@ -103,9 +101,9 @@ type Frontend struct {
 }
 
 type ExternalLink struct {
-	Text  string `json:"text"`
+	Text  string `json:"text" validate:"required"`
 	Title string `json:"title"`
-	Url   string `json:"url"`
+	Url   string `json:"url" validate:"required"`
 }
 
 // UserDefaults is a type that holds the default values
@@ -115,7 +113,6 @@ type UserDefaults struct {
 	DarkMode             bool                `json:"darkMode"`
 	LockPassword         bool                `json:"lockPassword"`
 	DisableSettings      bool                `json:"disableSettings,omitempty"`
-	Scope                string              `json:"scope"` // deprecated
 	Locale               string              `json:"locale"`
 	ViewMode             string              `json:"viewMode"`
 	GallerySize          int                 `json:"gallerySize"`