+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file. For commit 
    - some light value validation on certain fields
    - removed recaptcha -- was disabled and not used before.
    - moved `recaptcha` and `signup` configs to `auth.methods.password`
+ - also pushes to ghcr `ghcr.io/gtsteffaniak/filebrowser` https://github.com/gtsteffaniak/filebrowser/issues/470
 
  **BugFixes**:
  - fix scope reset on restart https://github.com/gtsteffaniak/filebrowser/issues/515